Fix directory tree
