+++ EbookWindows/EbookWindows.csproj
@@ -113,6 +113,7 @@
     <Compile Include="ViewModels\Book_Short_ViewModel.cs" />
     <Compile Include="ViewModels\Book_ViewModel.cs" />
     <Compile Include="ViewModels\Chapter_ViewModel.cs" />
+    <Compile Include="ViewModels\RecentFile_ViewModel.cs" />
     <Compile Include="View\PopupEbookScreen.xaml.cs">
       <DependentUpon>PopupEbookScreen.xaml</DependentUpon>
     </Compile>
@@ -139,12 +140,12 @@
     </Compile>
     <Compile Include="ViewModels\Converter.cs" />
     <Compile Include="ViewModels\myEpubReader.cs" />
-    <Compile Include="ViewModels\pdfDocumentViewModel.cs" />
+    <Compile Include="Model\pdfDocument.cs" />
     <Page Include="View\PopupEbookScreen.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Compile Include="ViewModels\RecentFile.cs" />
+    <Compile Include="Model\RecentFile.cs" />
     <Compile Include="Model\Chapter_Content.cs" />
     <Page Include="View\DetailScreen.xaml">
       <SubType>Designer</SubType>
