+++ EbookWindows/ViewModels/RecentFile_ViewModel.cs
@@ -5,35 +5,10 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows;
 
 namespace EbookWindows.ViewModels
 {
-    class RecentFile
-    {
-        public string fileName { get; set; }
-        public string filePath { get; set; }
-        public string fileIcon { get; set; }
-        public string recentLocation { get; set; }
-        public RecentFile() { }
-        public RecentFile(string fileName, string filePath, string fileIcon)
-        {
-            this.fileName = fileName;
-            this.filePath = filePath;
-            this.fileIcon = fileIcon;
-            this.recentLocation = "0";
-        }
-
-        public RecentFile(string fileName, string filePath, string fileIcon, string recentLocal)
-        {
-            this.fileName = fileName;
-            this.filePath = filePath;
-            this.fileIcon = fileIcon;
-            this.recentLocation = recentLocal;
-        }
-    }
-
-    class RecentFileDao
+    class RecentFile_ViewModel
     {
 
         //ObservableCollection
@@ -96,7 +71,7 @@ namespace EbookWindows.ViewModels
             //}
             StreamWriter fileStream = new StreamWriter(strFileName);
             int amount = recentFileList.Count();
-            if(amount > 10)
+            if (amount > 10)
             {
                 amount = 10;
             }
